[calibration] checker detection IO: move implementation to cpp file
